Split recipe pipeline modules
